Reorganized systems for cities
